+++ src/components/JobCard/index.tsx
@@ -1,47 +1,1 @@
-import Image from "next/image";
-import StackBadge from "src/components/StackBadge";
-import styles from "./JobCard.module.css";
-
-interface JobCardProps {
-  cardData: {
-    company: {
-      logo: string;
-      name: string;
-    };
-    job: {
-      title: string;
-      place: string;
-    };
-
-    stack: string[];
-  };
-}
-
-const JobCard = ({ cardData }: JobCardProps) => {
-  const { company, job, stack } = cardData;
-  const { logo, name } = company;
-  const { title, place } = job;
-
-  return (
-    <div id={styles.root}>
-      <div id={styles.content}>
-        <div id={styles.logoTitle}>
-          <img id={styles.logo} src={logo} />
-          <div id={styles.title}>
-            <h4 id={styles.name}>{name}</h4>
-            <h3 id={styles.job}>{title}</h3>
-          </div>
-        </div>
-        <h5 id={styles.place}>{place}</h5>
-      </div>
-      <hr />
-      <div id={styles.stack}>
-        {stack.map((tec) => (
-          <StackBadge name={tec} key={tec} />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export default JobCard;
+export { default } from './JobCard';
